moved listener tests to their own file
